v0810
Still doing QA task;
needs to review Read_Premissions and ReadByGroup actions because they
function locally but not in the server(deployment).  When Create or
Update in a Grid the SetPermissions DataAnnotation function 50% because
the user can not persist the data in DB but the sistem does not redirect
the user to the loggin page.